+++ package.json
@@ -126,8 +126,11 @@
     },
     "homepage": "https://github.com/atomicojs/atomico#readme",
     "devDependencies": {
+        "@vitest/browser": "^2.1.1",
+        "construct-style-sheets-polyfill": "^3.1.0",
         "happy-dom": "^15.7.4",
         "jsdom": "^25.0.0",
+        "playwright": "^1.47.2",
         "typescript": "^5.3.3"
     },
     "prettier": {
