+++ vite.config.js
@@ -3,6 +3,10 @@ import { defineConfig } from "vite";
 export default defineConfig({
     test: {
         include: ["./src/**/*.test.tsx"],
-        environment: "jsdom"
+        browser: {
+            provider: "playwright", // or 'webdriverio'
+            enabled: true,
+            name: "chromium" // browser name is required
+        }
     }
 });
